Replace http codes
