+++ package.json
@@ -3,12 +3,12 @@
   "version": "1.0.0",
   "type": "module",
   "description": "Типотека",
-  "main": "src/api/service.js",
+  "main": "src/cli/main.js",
   "scripts": {
     "test": "npm run eslint && npm run jest",
     "jest": "jest",
     "eslint": "eslint --ext .js src/",
-    "start": "nodemon src/api/service.js",
+    "start": "nodemon src/cli/main.js",
     "start-web-app": "node src/web-app/main.js",
     "generate-mocks": "node src/mock-generator/index.js 5 6 mocks.json",
     "dev-server": "nodemon src/web-app/main.js"
