+++ src/api/article/article-router.js
@@ -1,36 +1,34 @@
-'use strict';
-
+import {constants as HTTP_CODES} from 'http2';
 import {Router} from 'express';
-import {HTTP_CODE} from '../../../constants';
-import {validateArticle} from './validate-article';
-import {validateComment} from './validate-comment';
+import {validateArticle} from './validate-article.js';
+import {validateComment} from './validate-comment.js';
 
 export function createArticleRouter(app, articleService) {
   const articlesRouter = new Router();
 
   articlesRouter.get(`/`, (req, res) => {
     const articles = articleService.findAll();
-    res.status(HTTP_CODE.OK);
+    res.status(HTTP_CODES.HTTP_STATUS_OK);
     res.json(articles);
   });
 
   articlesRouter.get(`/:articleId`, (req, res) => {
     const articleId = req.params.articleId;
     const article = articleService.findById(articleId);
-    res.status(HTTP_CODE.OK);
+    res.status(HTTP_CODES.HTTP_STATUS_OK);
     res.json(article);
   });
 
   articlesRouter.get(`/:articleId/comments`, (req, res) => {
     const articleId = req.params.articleId;
     const comments = articleService.findCommentsByArticleId(articleId);
-    res.status(HTTP_CODE.OK);
+    res.status(HTTP_CODES.HTTP_STATUS_OK);
     res.json(comments);
   });
 
   articlesRouter.post(`/`, validateArticle, (req, res) => {
     articleService.create(req.body);
-    res.status(HTTP_CODE.OK);
+    res.status(HTTP_CODES.HTTP_STATUS_OK);
     res.end();
   })
 
@@ -38,14 +36,14 @@ export function createArticleRouter(app, articleService) {
     const articleId = req.params.articleId;
     const articleData = req.body;
     articleService.update(articleId, articleData);
-    res.status(HTTP_CODE.OK);
+    res.status(HTTP_CODES.HTTP_STATUS_OK);
     res.end();
   });
 
   articlesRouter.delete(`/:articleId`, (req, res) => {
     const articleId = req.params.articleId;
     articleService.delete(articleId);
-    res.status(HTTP_CODE.OK);
+    res.status(HTTP_CODES.HTTP_STATUS_OK);
     res.end();
   });
 
@@ -53,7 +51,7 @@ export function createArticleRouter(app, articleService) {
     const articleId = req.params.articleId;
     const commentId = req.params.commentId;
     articleService.deleteCommentById(articleId, commentId);
-    res.status(HTTP_CODE.OK);
+    res.status(HTTP_CODES.HTTP_STATUS_OK);
     res.end();
   });
 
@@ -61,7 +59,7 @@ export function createArticleRouter(app, articleService) {
     const articleId = req.params.articleId;
     const commentData = req.body;
     articleService.addComment(articleId, commentData);
-    res.status(HTTP_CODE.OK);
+    res.status(HTTP_CODES.HTTP_STATUS_OK);
     res.end();
   })
