+++ src/api/category/category-router.js
@@ -1,14 +1,12 @@
-'use strict';
-
 import {Router} from 'express';
-import {HTTP_CODE} from '../../../constants.js';
+import {constants as HTTP_CODES} from 'http2';
 
 export function createCategoryRouter(app, categoryService) {
   const categoriesRouter = new Router();
 
   categoriesRouter.get(`/`, async (req, res) => {
     const categories = await categoryService.findAll();
-    res.status(HTTP_CODE.OK);
+    res.status(HTTP_CODES.HTTP_STATUS_OK);
     res.json(categories);
   });
