+++ src/api/search/search-router.js
@@ -1,7 +1,5 @@
-'use strict';
-
 import {Router} from 'express';
-import {HTTP_CODE} from '../../../constants.js';
+import {constants as HTTP_CODES} from 'http2';
 
 export function createSearchRouter(app, searchService) {
   const searchRouter = new Router();
@@ -9,7 +7,7 @@ export function createSearchRouter(app, searchService) {
   searchRouter.get(`/`, (req, res) => {
     const searchQuery = req.query.query;
     const result = searchService.searchFor(searchQuery);
-    res.status(HTTP_CODE.OK);
+    res.status(HTTP_CODES.HTTP_STATUS_OK);
     res.json(result);
   });
