+++ src/cli/Cli.js
@@ -1,0 +1,13 @@
+'use strict';
+
+import {versionCommand} from './commands/version.js';
+import {helpCommand} from './commands/help.js';
+import {generateMocksCommand} from './commands/generate-mocks.js';
+import {runApiCommand} from './commands/run-api.js';
+
+export const Cli = {
+  [versionCommand.name]: versionCommand,
+  [helpCommand.name]: helpCommand,
+  [generateMocksCommand.name]: generateMocksCommand,
+  [runApiCommand.name]: runApiCommand,
+};
