+++ src/cli/commands/generate-mocks.js
