+++ src/cli/commands/help.js
