+++ src/cli/commands/version.js
